shopping_cart/requirements.txt
fastapi==0.111.1
uvicorn==0.24.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
sqlalchemy==2.0.19
pydantic==2.5.2
requests ==2.32.5
pytest==8.4.2